Add error stream option to RLPGatewayClient
This allows users of the client to be notified of errors.

The biggest limitation of the previous implementation is if the UAA
token expires, the client will go into a hot loop trying to reconnect
and the calling code has no way to react to it except for hacky log
processing.

This also adds a one second delay between reconnects to reduce CPU load
and load on the RLP Gateway itself.