Migrate golangci-lint cfg to v2. Fix new findings
The new v2 version of `golangci-lint` comes with a new configuration file
structure and therefore the configuration had to be migrated.
`golangci-lint migrate` had to be executed and copy the comments to
the new config as they are not touched by the migration. There were also
few minor new findings that had to be fixed.